make delete comment for author

+++ src/Controller/CommentController.php
@@ -17,11 +17,16 @@ class CommentController extends AbstractController
     #[Route('/{id}', name: 'app_comment_delete', methods: ['POST'])]
     public function delete(Request $request, Comment $comment, EntityManagerInterface $entityManager): Response
     {
+        if (!$this->getUser() || $this->getUser() !== $comment->getUser()) {
+            return $this->redirectToRoute('app_home');
+        }
+        
         if ($this->isCsrfTokenValid('delete'.$comment->getId(), $request->request->get('_token'))) {
             $entityManager->remove($comment);
             $entityManager->flush();
         }
-
-        return $this->redirectToRoute('app_comment_index', [], Response::HTTP_SEE_OTHER);
+        
+        $targetUrl = $this->generateUrl('app_trick_show', ['id' => $comment->getTrick()->getId()]) . '#commentSection';
+        return $this->redirect($targetUrl);
     }
 }
